pyproject.toml
[project]
name = "youtube-2-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=5.3.2",
    "fastapi>=0.128.1",
    "mistralai>=1.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.1.31",
]
